requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.16.5
    # via currency-converter (pyproject.toml)
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
currencyapicom==0.1.1
    # via currency-converter (pyproject.toml)
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
everapi==0.1.1
    # via currencyapicom
fastapi==0.118.0
    # via currency-converter (pyproject.toml)
fastapi-cli==0.0.13
    # via fastapi
fastapi-cloud-cli==0.2.1
    # via fastapi-cli
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via fastapi
jinja2==3.1.6
    # via fastapi
loguru==0.7.3
    # via currency-converter (pyproject.toml)
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
orjson==3.11.3
    # via fastapi
packaging==25.0
    # via pytest
pluggy==1.6.0
    # via pytest
psycopg2-binary==2.9.10
    # via currency-converter (pyproject.toml)
pydantic==2.11.9
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-extra-types==2.10.5
    # via fastapi
pydantic-settings==2.11.0
    # via fastapi
pygments==2.19.2
    # via
    #   pytest
    #   rich
pytest==8.4.2
    # via
    #   currency-converter (pyproject.toml)
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via currency-converter (pyproject.toml)
python-dotenv==1.1.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.3
    # via
    #   fastapi
    #   uvicorn
redis==6.4.0
    # via currency-converter (pyproject.toml)
requests==2.32.5
    # via
    #   currencyapicom
    #   everapi
rich==14.1.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
ruff==0.13.2
    # via currency-converter (pyproject.toml)
sentry-sdk==2.39.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.43
    # via
    #   currency-converter (pyproject.toml)
    #   alembic
starlette==0.48.0
    # via fastapi
typer==0.19.2
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   pytest-asyncio
    #   rich-toolkit
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
ujson==5.11.0
    # via fastapi
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.37.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
